refactor(core): typed canonical value vocabulary between source and sink
GenericValue collapsed uuid/date/timestamp/time to String, so a value
crossed the source→sink boundary type-erased — fine for OpenSearch
(which reads the field's mapping), but the next sink can't tell a date
from a string, and re-binding a key into SQL hit operator-does-not-exist.

Make GenericValue the typed middle type both sides translate through:
- Split numerics by width (SmallInt/Int/BigInt, Float/Double, Decimal),
  split temporals (Date/Time/Timestamp/TimestampTz), add Uuid and Bytes.
  Nothing semantic is erased in transit.
- Serde is now derived and format-agnostic (externally tagged), so a
  value round-trips losslessly through any queue encoding — in as a
  GenericValue, out as the same GenericValue. Core picks no wire format;
  hand-rolled Eq/Hash remain only because f32/f64 need bit-equality.
- The Postgres source decodes each column into its typed variant and
  binds typed values natively; the batched lookup selects its key column
  raw so the returned key matches the requested typed key exactly.
- The sink boundary (sinks_core::to_json) renders the typed variants to
  the JSON OpenSearch ingests — temporals as ISO strings, uuid as its
  string, bytea as base64. The base64 lives at the sink, not in core.

Config literals (constant/default/options/extra) stay GenericValue,
round-tripping through the lock via the derived tagged serde.

+++ libs/1-sinks/2-opensearch/tests/reindex.rs
@@ -120,7 +120,7 @@ fn mapping() -> IndexMapping {
 
 fn document(i: usize) -> GenericValue {
     let mut map = BTreeMap::new();
-    map.insert("id".to_owned(), GenericValue::Int(i as i64));
+    map.insert("id".to_owned(), GenericValue::BigInt(i as i64));
     map.insert(
         "email".to_owned(),
         GenericValue::String(format!("user{i}@example.com")),
